+++ apiron/client.py
@@ -109,6 +109,7 @@ class ServiceCaller:
         data=None,
         headers=None,
         cookies=None,
+        auth=None,
     ):
         host = cls.choose_host(service=service)
 
@@ -127,6 +128,7 @@ class ServiceCaller:
             data=data,
             headers=headers,
             cookies=cookies,
+            auth=auth,
         )
 
         return session.prepare_request(request)
@@ -143,6 +145,8 @@ class ServiceCaller:
         data=None,
         headers=None,
         cookies=None,
+        auth=None,
+        encoding=None,
         retry_spec=DEFAULT_RETRY,
         timeout_spec=DEFAULT_TIMEOUT,
         logger=None,
@@ -176,6 +180,12 @@ class ServiceCaller:
             (default ``None``)
         :param dict cookies:
             Cookies to send to the endpoint
+            (default ``None``)
+        :param auth:
+            An object suitable for the :class:`requests.Request` object's ``auth`` argument
+        :param str encoding:
+            The codec to use when decoding the response.
+            Default behavior is to have ``requests`` guess the codec.
             (default ``None``)
         :param urllib3.util.retry.Retry retry_spec:
             (optional)
@@ -214,6 +224,7 @@ class ServiceCaller:
             data=data,
             headers=headers,
             cookies=cookies,
+            auth=auth,
         )
 
         logger.info('{method} {url}'.format(
@@ -237,5 +248,8 @@ class ServiceCaller:
             session.close()
 
         response.raise_for_status()
+
+        if encoding:
+            response.encoding = encoding
 
         return cls.choose_response_format(endpoint, response, return_raw_response)