+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Ability to expose raw response returned by `requests` when using `apiron.client.ServiceCaller.call`
 
+## [1.1.0] - 2018-12-03
+### Added
+- Exposed ability to explicitly set the response encoding
+- Expose pass-through `auth` argument to `requests`
+
 ### Changed
 - Moved all tests to a top-level `tests` package and switched to using the `pytest` runner
 - Added missing parameter in docstring for `apiron.client.ServiceCaller.call`